+++ romsearch/util/__init__.py
@@ -1,6 +1,6 @@
 from .discord import discord_push
 from .general import split, match_retool_search_terms, normalize_name, get_file_time
-from .io import load_yml, save_yml, unzip_file, load_json, save_json
+from .io import load_yml, save_yml, unzip_file, load_json, save_json, get_dat, format_dat
 from .logger import setup_logger, centred_string, left_aligned_string
 from .regex_matching import (
     get_file_pattern,
@@ -31,4 +31,6 @@ __all__ = [
     "get_file_time",
     "normalize_name",
     "match_retool_search_terms",
+    "get_dat",
+    "format_dat",
 ]